Fix tag creation
Previously, the tag creation step did not actually push the tag that
was created to remote. When the release step ran, it created a new
tag using the default target (master). Although a new release was still
created with the right assets, it caused problems with how GitHub sorts
releases.

GitHub releases seem to be sorted according to the following rules:

1. If the release name starts with a v ie is a semver version, it will
   appear first. Releases are sorted in descending order ie v3, v2, v1.
2. Use the tag date or commit date (depending on whether its a
   lightweight tag or annotated tag). Sort in descending order ie newest
   first. If there are multiple releases/tags on the same day, use the
   name sorted in descending string order ie c, b, a.

The following command can be used to confirm the order in the GitHub
releases page

git for-each-ref --sort="-refname" --sort="-committerdate" --format '%(committerdate) %(refname:short)' refs/tags

Since we were tagging the master branch repeatedly, this resulted in
multiple tags sharing the same commit and therefore commit date. This
meant that GitHub fell back to using the tag/release name which is
based on the Etag returned from nvd. Since it is a hash, it did not
give a predictable order when sorting. This would be resolved every time
the keep alive workflow ran, because a new commit on the default branch
meant that any new tags that were created pointed to the new commit
which had a newer date. Then, the issue would happen again when the
daily generate workflow ran a 2nd time after.